Remove old SearchDropdown component from BorrowTokenPage Add SearchInput to ModalTokenSelect
